# project-docs/openapi.yaml
openapi: 3.0.3
info:
  title: Forge API
  description: |
    REST API for Forge, a file-system-backed project management tool for software teams.
    All request/response bodies use **snake_case** field names.
    Paginated list endpoints return `{ items, total, page, limit, total_pages }`.
  version: 1.0.0
servers:
  - url: http://localhost:3001
    description: Local development server

# ─────────────────────────────────────────────────────────────────────────────
# Paths
# ─────────────────────────────────────────────────────────────────────────────
paths:
...
  /api/issues:
    get:
      summary: List issues (paginated)
      operationId: listIssues
      tags: [Issues]
      parameters:
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/limit"
        - name: status
          in: query
          schema: { type: string }
          description: Comma-separated statuses (open, in-progress, closed)
        - name: type
          in: query
          schema: { type: string }
          description: Comma-separated types (bug, task, improvement, chore)
        - name: priority
          in: query
          schema: { type: string }
          description: Comma-separated priorities (critical, high, medium, low)
        - name: assignee
          in: query
          schema: { type: string }
          description: Filter by assignee or `__unassigned__`
        - name: feature
          in: query
          schema: { type: string }
          description: Filter by linked feature slug or `__unlinked__`
        - name: points_min
          in: query
          schema: { type: number }
        - name: points_max
          in: query
          schema: { type: number }
        - name: labels
          in: query
          schema: { type: string }
          description: Comma-separated labels
        - name: sort
          in: query
          schema:
            type: string
            enum: [order, title, status, type, priority, modified]
            default: order
        - name: dir
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: asc
      responses:
        "200":
          description: Paginated list of issues
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedIssues"

    post:
      summary: Create an issue
      operationId: createIssue
      tags: [Issues]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIssueInput"
      responses:
        "201":
          description: Issue created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Issue"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/issues/reorder:
    patch:
      summary: Reorder issues
      operationId: reorderIssues
      tags: [Issues]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReorderInput"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/issues/{slug}:
    parameters:
      - $ref: "#/components/parameters/slug"

    get:
      summary: Get an issue by slug
      operationId: getIssue
      tags: [Issues]
      responses:
        "200":
          description: Issue detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Issue"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      summary: Update an issue
      operationId: updateIssue
      tags: [Issues]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateIssueInput"
      responses:
        "200":
          description: Updated issue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Issue"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"

    delete:
      summary: Delete an issue
      operationId: deleteIssue
      tags: [Issues]
      responses:
        "200":
          description: Issue deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  slug: { type: string }
        "404":
          $ref: "#/components/responses/NotFound"
  /api/issues/{slug}/artifacts:
    parameters:
      - $ref: "#/components/parameters/slug"
    get:
      summary: List artifacts for an issue
      operationId: listIssueArtifacts
      tags: [Issues]
      responses:
        "200":
          description: List of artifacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Artifact"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/issues/{slug}/artifacts/upload:
    parameters:
      - $ref: "#/components/parameters/slug"
    post:
      summary: Upload a file artifact to an issue
      operationId: uploadIssueArtifact
      tags: [Issues]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required: [file]
      responses:
        "201":
          description: Artifact uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artifact"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/issues/{slug}/artifacts/create:
    parameters:
      - $ref: "#/components/parameters/slug"
    post:
      summary: Create an empty text artifact for an issue
      operationId: createIssueArtifact
      tags: [Issues]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateArtifactInput"
      responses:
        "201":
          description: Artifact created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artifact"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/issues/{slug}/artifacts/{filename}:
    parameters:
      - $ref: "#/components/parameters/slug"
      - $ref: "#/components/parameters/filename"

    get:
      summary: Get issue artifact content
      operationId: getIssueArtifactContent
      tags: [Issues]
      parameters:
        - name: raw
          in: query
          schema: { type: boolean }
          description: Force raw binary response
      responses:
        "200":
          description: Artifact content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtifactContent"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      summary: Save issue artifact content
      operationId: saveIssueArtifactContent
      tags: [Issues]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveArtifactContentInput"
      responses:
        "200":
          description: Artifact saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artifact"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"

    delete:
      summary: Delete an issue artifact
      operationId: deleteIssueArtifact
      tags: [Issues]
      responses:
        "200":
          description: Artifact deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  slug: { type: string }
                  filename: { type: string }
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Wiki ──────────────────────────────────────────────────────────────────
...
components:
...
  parameters:
    page:
      name: page
      in: query
      schema: { type: integer, default: 1, minimum: 1 }
      description: Page number
    limit:
      name: limit
      in: query
      schema: { type: integer, default: 25, minimum: 1, maximum: 100 }
      description: Items per page
    slug:
      name: slug
      in: path
      required: true
      schema: { type: string }
    filename:
      name: filename
      in: path
      required: true
      schema: { type: string }

  # ── Responses ─────────────────────────────────────────────────────────────
  responses:
    Ok:
      description: Success
      content:
        application/json:
          schema:
            type: object
            properties:
              ok: { type: boolean, example: true }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }
    ValidationError:
      description: Validation error (422)
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string, description: "Semicolon-separated list of field-level errors" }
    Conflict:
      description: Resource already exists (409)
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }

  # ── Schemas ───────────────────────────────────────────────────────────────
  schemas:
...
    IssueStatus:
      type: string
      enum: [open, in-progress, closed]
    IssueType:
      type: string
      enum: [bug, task, improvement, chore]
    Priority:
      type: string
      enum: [critical, high, medium, low]

    # ── Feature ────────────────────────────────────────────────────────────
...
    Issue:
      type: object
      properties:
        slug: { type: string }
        title: { type: string }
        status: { $ref: "#/components/schemas/IssueStatus" }
        type: { $ref: "#/components/schemas/IssueType" }
        priority: { $ref: "#/components/schemas/Priority" }
        assignees:
          type: array
          items: { type: string }
        points: { type: number, nullable: true }
        sprint: { type: string, nullable: true }
        feature: { type: string, nullable: true, description: Linked feature slug }
        labels:
          type: array
          items: { type: string }
        created: { type: string }
        modified: { type: string }
        body: { type: string }
        order: { type: number }
    PaginatedIssues:
      type: object
      properties:
        issues:
          type: array
          items: { $ref: "#/components/schemas/Issue" }
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
        total_pages: { type: integer }
    CreateIssueInput:
      type: object
      required: [title]
      properties:
        title: { type: string, minLength: 1 }
        body: { type: string }
        type: { $ref: "#/components/schemas/IssueType" }
        priority: { $ref: "#/components/schemas/Priority" }
        feature: { type: string }
    UpdateIssueInput:
      type: object
      properties:
        title: { type: string, minLength: 1 }
        status: { $ref: "#/components/schemas/IssueStatus" }
        type: { $ref: "#/components/schemas/IssueType" }
        priority: { $ref: "#/components/schemas/Priority" }
        assignees:
          type: array
          items: { type: string }
        points: { type: number, nullable: true }
        sprint: { type: string, nullable: true }
        feature: { type: string, nullable: true }
        labels:
          type: array
          items: { type: string }
        body: { type: string }

    # ── Wiki ───────────────────────────────────────────────────────────────
...
    ReorderInput:
      type: object
      required: [slugs]
      properties:
        slugs:
          type: array
          items: { type: string }
        offset: { type: number, description: Starting offset for the order values }

    # ── Artifacts ──────────────────────────────────────────────────────────
    Artifact:
      type: object
      properties:
        name: { type: string }
        size: { type: number }
        type: { type: string }
    ArtifactContent:
      type: object
      properties:
        name: { type: string }
        content: { type: string }
        type: { type: string }
        mime_type: { type: string }
        size: { type: number }
    CreateArtifactInput:
      type: object
      required: [filename]
      properties:
        filename: { type: string, minLength: 1 }
    SaveArtifactContentInput:
      type: object
      required: [content]
      properties:
        content: { type: string }

    # ── Search ─────────────────────────────────────────────────────────────
...